clean up new entity modal #36
